Create Portfolio
My personal portfolio using HTML, CSS, and JavaScript. This took a few days to complete, showcasing my proficiency in web development.

I will be updating this as I start creating and finishing websites to showcase on here. As of right now, these are projects I want to start working on.